refactor: Create models.deployment package and move all related deployment classes there
