+++ library/src/main/java/com/avioconsulting/mule/deployment/api/models/deployment/ExchangeAppDeploymentRequest.groovy
@@ -1,4 +1,4 @@
-package com.avioconsulting.mule.deployment.api.models
+package com.avioconsulting.mule.deployment.api.models.deployment
 
 abstract class ExchangeAppDeploymentRequest extends AppDeploymentRequest {
