+++ library/src/main/java/com/avioconsulting/mule/deployment/api/models/deployment/RuntimeFabricDeploymentRequest.groovy
@@ -1,5 +1,7 @@
-package com.avioconsulting.mule.deployment.api.models
+package com.avioconsulting.mule.deployment.api.models.deployment
 
+
+import com.avioconsulting.mule.deployment.api.models.WorkerSpecRequest
 import com.avioconsulting.mule.deployment.internal.models.CloudhubAppProperties
 import groovy.json.JsonOutput
 import groovy.transform.ToString
