+++ library/src/main/java/com/avioconsulting/mule/deployment/dsl/ApiSpecContext.groovy
@@ -1,7 +1,7 @@
 package com.avioconsulting.mule.deployment.dsl
 
 import com.avioconsulting.mule.deployment.api.models.ApiSpecification
-import com.avioconsulting.mule.deployment.api.models.FileBasedAppDeploymentRequest
+import com.avioconsulting.mule.deployment.api.models.deployment.FileBasedAppDeploymentRequest
 
 class ApiSpecContext extends BaseContext {
     String name, exchangeAssetId, mainRamlFile, endpoint, autoDiscoveryPropertyName, designCenterBranchName, sourceDirectory, soapEndpointWithVersion
