+++ library/src/main/java/com/avioconsulting/mule/deployment/dsl/CloudhubV2Context.groovy
@@ -1,6 +1,6 @@
 package com.avioconsulting.mule.deployment.dsl
 
-import com.avioconsulting.mule.deployment.api.models.CloudhubV2DeploymentRequest
+import com.avioconsulting.mule.deployment.api.models.deployment.CloudhubV2DeploymentRequest
 
 class CloudhubV2Context extends RuntimeFabricContext {
