+++ library/src/main/java/com/avioconsulting/mule/deployment/dsl/DeploymentPackage.groovy
@@ -2,9 +2,8 @@ package com.avioconsulting.mule.deployment.dsl
 
 
 import com.avioconsulting.mule.deployment.api.models.ApiSpecificationList
-import com.avioconsulting.mule.deployment.api.models.AppDeploymentRequest
+import com.avioconsulting.mule.deployment.api.models.deployment.AppDeploymentRequest
 import com.avioconsulting.mule.deployment.api.models.Features
-import com.avioconsulting.mule.deployment.api.models.FileBasedAppDeploymentRequest
 import com.avioconsulting.mule.deployment.api.models.policies.Policy
 import groovy.transform.Canonical
