+++ library/src/main/java/com/avioconsulting/mule/deployment/internal/subdeployers/ICloudHubDeployer.groovy
@@ -1,6 +1,6 @@
 package com.avioconsulting.mule.deployment.internal.subdeployers
 
-import com.avioconsulting.mule.deployment.api.models.CloudhubDeploymentRequest
+import com.avioconsulting.mule.deployment.api.models.deployment.CloudhubDeploymentRequest
 
 interface ICloudHubDeployer extends ISubDeployer<CloudhubDeploymentRequest> {
     def deploy(CloudhubDeploymentRequest deploymentRequest)
