+++ library/src/main/java/com/avioconsulting/mule/deployment/internal/subdeployers/IDesignCenterDeployer.groovy
@@ -1,7 +1,7 @@
 package com.avioconsulting.mule.deployment.internal.subdeployers
 
 import com.avioconsulting.mule.deployment.api.models.ApiSpecification
-import com.avioconsulting.mule.deployment.api.models.FileBasedAppDeploymentRequest
+import com.avioconsulting.mule.deployment.api.models.deployment.FileBasedAppDeploymentRequest
 
 interface IDesignCenterDeployer {
     def synchronizeDesignCenterFromApp(ApiSpecification apiSpec,
