+++ library/src/main/java/com/avioconsulting/mule/deployment/internal/subdeployers/IOnPremDeployer.groovy
@@ -1,6 +1,6 @@
 package com.avioconsulting.mule.deployment.internal.subdeployers
 
-import com.avioconsulting.mule.deployment.api.models.OnPremDeploymentRequest
+import com.avioconsulting.mule.deployment.api.models.deployment.OnPremDeploymentRequest
 
 interface IOnPremDeployer extends ISubDeployer<OnPremDeploymentRequest> {
     def deploy(OnPremDeploymentRequest deploymentRequest)
