+++ library/src/main/java/com/avioconsulting/mule/deployment/internal/subdeployers/IRuntimeFabricDeployer.groovy
@@ -1,6 +1,6 @@
 package com.avioconsulting.mule.deployment.internal.subdeployers
 
-import com.avioconsulting.mule.deployment.api.models.RuntimeFabricDeploymentRequest
+import com.avioconsulting.mule.deployment.api.models.deployment.RuntimeFabricDeploymentRequest
 
 interface IRuntimeFabricDeployer extends ISubDeployer<RuntimeFabricDeploymentRequest> {
     def deploy(RuntimeFabricDeploymentRequest deploymentRequest)
