+++ library/src/main/java/com/avioconsulting/mule/deployment/internal/subdeployers/ISubDeployer.groovy
@@ -1,6 +1,6 @@
 package com.avioconsulting.mule.deployment.internal.subdeployers
 
-import com.avioconsulting.mule.deployment.api.models.AppDeploymentRequest
+import com.avioconsulting.mule.deployment.api.models.deployment.AppDeploymentRequest
 
 interface ISubDeployer<T extends AppDeploymentRequest> {
     boolean isMule4Request(T deploymentRequest)
