+++ library/src/test/java/com/avioconsulting/mule/deployment/api/models/CloudhubDeploymentRequestTest.groovy
@@ -1,6 +1,7 @@
 package com.avioconsulting.mule.deployment.api.models
 
 import com.avioconsulting.mule.MavenInvoke
+import com.avioconsulting.mule.deployment.api.models.deployment.CloudhubDeploymentRequest
 import org.junit.BeforeClass
 import org.junit.Test
